+++ spicedmoon/basics.py
@@ -1,20 +1,22 @@
+"""
+Common basic functions that help and improve in SPICE usage.
+"""
 import os
 import time
 from typing import List, Union, Tuple
 from datetime import datetime, timezone
+import warnings
 
 import numpy as np
 import spiceypy as spice
 
-from .constants import CUSTOM_KERNEL_NAME
+from .constants import CUSTOM_KERNEL_NAME, MOON_EQ_RAD, MOON_POL_RAD
 
 
 def furnsh_safer(k_path: str):
     """
     Performs SPICE's `furnsh_c`, but in case that it fails it tries again after a small time
-    interval.
-
-    Furnsh very rarely crashes, but it can be solved trying again.
+    interval. Furnsh very rarely crashes, but it can be solved trying again.
 
     Parameters
     ----------
@@ -28,24 +30,42 @@ def furnsh_safer(k_path: str):
         spice.furnsh(k_path)
 
 
+def _is_dt_tz_aware(dt: datetime) -> bool:
+    """Checks if a datetime is timezone aware or not
+
+    Parameters
+    ----------
+    dt: datetime
+        Datetime to check timezone-awareness
+
+    Result
+    ------
+    is_aware: bool
+        Timezone-awareness of the datetime
+    """
+    return dt.tzinfo is None or dt.tzinfo.utcoffset(dt) is None
+
+
 def dt_to_str(dts: Union[List[datetime], List[str]]) -> List[str]:
-    """Convert a list of datetimes into a list of string dates in a valid format.
+    """Convert a list of datetimes into a list of string dates in a valid SPICE `str` format.
 
     Parameters
     ----------
     dts: list of datetimes | list of str
-        List of datetimes that will be converted to utc_times. They must be timezone aware.
-        A list of already converted strings can be given instead, and it will be returned without
+        List of datetimes that will be converted to `str`. They must be timezone aware.
+        A list of already `str` can be given instead, and it will be returned without
         change.
 
     Returns
     -------
     utc_times: list of str
-        List of the datetimes in a valid string format for SPICE.
+        List of the timestamps in a valid `str` format for SPICE.
     """
     utc_times = []
     for dt in dts:
         if isinstance(dt, datetime):
+            if not _is_dt_tz_aware(dt):
+                warnings.warn("Using timezone-naive datetime object", RuntimeWarning)
             dt_utc = dt.astimezone(timezone.utc)
             utc_times.append(dt_utc.strftime("%Y-%m-%d %H:%M:%S"))
         else:
@@ -83,9 +103,9 @@ def get_radii_moon(ignore_bodvrd: bool = True) -> Tuple[float, float]:
     pol_rad: float
         Polar radius of the Moon.
     """
-    eq_rad, pol_rad = 1738.1, 1736
+    eq_rad, pol_rad = MOON_EQ_RAD, MOON_POL_RAD
     if not ignore_bodvrd:
-        # The ones obtained with bodvrd are not correct
+        # The ones obtained with bodvrd are not correct, not accurate
         _, radii_moon = spice.bodvrd("MOON", "RADII", 3)
         eq_rad, pol_rad = radii_moon[0], radii_moon[2]
     return eq_rad, pol_rad